requirements.txt
nltk == 3.3
progressbar2 == 3.37.1
scikit_learn[alldeps]
